Update RC button circular types
