+++ packages/components/src/__rc__/Button/Button.tsx
@@ -1,40 +1,10 @@
 import React, { forwardRef } from 'react'
-import { Button as RACButton, type ButtonProps as RACButtonProps } from 'react-aria-components'
+import { Button as RACButton } from 'react-aria-components'
 import { useReversedColors } from '~components/__utilities__/v3'
 import { mergeClassNames } from '~components/utils/mergeClassNames'
 import { ButtonContent, PendingContent } from './subcomponents'
-import { type ButtonSizes, type ButtonVariants, type PendingButtonProps } from './types'
+import { type ButtonProps, type PendingButtonProps } from './types'
 import styles from './Button.module.css'
-
-/** Shared UI props between Button and LinkButton */
-export type ButtonUIProps = {
-  /** Visually hides the Button's child content used as the label and the `pendingLabel`. Use for icon-only `Button`. @default "false" */
-  hasHiddenLabel?: boolean
-  /** The visual style of the button.
-   *  @default "default" */
-  variant?: ButtonVariants
-  /** The visual size of the button. `medium` was formerly `regular`
-   *  @default "medium" */
-  size?: ButtonSizes
-  /** Renders an icon at the `iconPosition` provided. To the size scales with the button, we recommend using the `Icon` component from `"@kaizen/components/future"` */
-  icon?: JSX.Element
-  /** Controls the position of the Icon passed in as props. @default "start" */
-  iconPosition?: 'start' | 'end'
-  /** Controls if the button inherits width from its parent. @default "false" */
-  isFullWidth?: boolean
-  /**
-   * Controls the reversed style of Button
-   * @deprecated Use the ReversedColors Provider instead. This is here to support gradual migration to the ReversedColors Provider and will take precedence if a value is provided. {@link https://cultureamp.design/?path=/docs/actions-button-button-v3-api-specification--docs#variants}
-   */
-  isReversed?: boolean
-}
-
-export type ButtonProps = ButtonUIProps &
-  PendingButtonProps &
-  Omit<RACButtonProps, 'children'> & {
-    /** Used as the label for the button. */
-    children: RACButtonProps['children']
-  }
 
 export const Button = forwardRef(
   (
